# Speed up `unchop()`