@hadley 2021-06-14
How close a review do you think this needs? i.e. do you have any worries that this might change behaviour in some edge case? If so, I probably won't be able to review it until I'm next work on tidyr, since I'll need to load it all up into my brain again.