@mgirlich 2021-06-15
There are quite a lot of changes, so a review would definitely make sense. But it's fine for me to wait until you have time to work again on tidyr.